refactor(Point) PointsTest->PointTest로 테스트코드 이동
